+++ src/v2/modules/invoice-account-addresses/routes.js
@@ -5,10 +5,12 @@ const controller = require('./controller');
 const entityHandlers = require('../../lib/entity-handlers');
 const validators = require('../../lib/validators');
 
+const getLocation = entity => `/crm/2.0/invoice-account-addresses/${entity.invoiceAccountAddressId}`;
+
 exports.postInvoiceAccountAddress = {
   method: 'POST',
   path: '/crm/2.0/invoice-accounts/{invoiceAccountId}/addresses',
-  handler: (request, h) => entityHandlers.createEntity(request, h, 'invoiceAccountAddress'),
+  handler: (request, h) => entityHandlers.createEntity(request, h, 'invoiceAccountAddress', getLocation),
   options: {
     description: 'Adds an address to an invoice account entity',
     validate: {
@@ -35,7 +37,6 @@ exports.deleteInvoiceAccountAddress = {
     description: 'Delete an invoice account address entity by id',
     validate: {
       params: {
-        invoiceAccountId: validators.GUID,
         invoiceAccountAddressId: validators.GUID
       }
     }
