+++ test/v2/modules/invoice-account-addresses/controller.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const { experiment } = exports.lab = require('@hapi/lab').script();
+
+experiment('v2/modules/invoice-account-addresses/controller', () => {
+
+});
