+++ test/v2/services/invoice-accounts.js
@@ -267,320 +267,10 @@ experiment('v2/services/invoice-accounts', () => {
     });
   });
 
-  experiment('.createInvoiceAccountAddress', () => {
-    beforeEach(async () => {
-      invoiceAccountsRepo.findOne.resolves(createInvoiceAccount());
-      addressesRepo.findOneWithCompanies.resolves({
-        companyAddresses: [{
-          companyId
-        }]
-      });
-      contactsRepo.findOneWithCompanies.resolves({
-        companyContacts: [{
-          companyId
-        }]
-      });
-    });
-
-    experiment('when the invoice account address data is invalid', () => {
-      let invoiceAccountAddress;
-      beforeEach(() => {
-        invoiceAccountAddress = {
-          invoiceAccountId: 'not-valid',
-          addressId: '123abc',
-          startDate: '2020-04-01',
-          agentCompanyId: null,
-          contactId: null
-        };
-      });
-
-      test('any EntityValidationError is thrown', async () => {
-        const err = await expect(invoiceAccountsService.createInvoiceAccountAddress(invoiceAccountAddress))
-          .to.reject(errors.EntityValidationError, 'Invoice account address not valid');
-
-        expect(err.validationDetails).to.be.an.array();
-        expect(err.message).to.equal('Invoice account address not valid');
-      });
-
-      test('the invoice account address is not saved', async () => {
-        expect(invoiceAccountAddressesRepo.create.called).to.equal(false);
-      });
-    });
-
-    experiment('when there is no agent company', () => {
-      let invoiceAccountAddress;
-
-      beforeEach(async () => {
-        invoiceAccountAddress = {
-          invoiceAccountId: uuid(),
-          addressId: uuid(),
-          startDate: '2020-04-01',
-          agentCompanyId: null,
-          contactId: null
-        };
-      });
-
-      experiment('and the posted address ID does not belong to the licence holder company', () => {
-        beforeEach(async () => {
-          addressesRepo.findOneWithCompanies.resolves({
-            companyAddresses: [{
-              companyId: 'some-other-id'
-            }]
-          });
-        });
-
-        test('a ConflictingDataError is thrown', async () => {
-          const err = await expect(invoiceAccountsService.createInvoiceAccountAddress(invoiceAccountAddress))
-            .to.reject(errors.ConflictingDataError);
-          expect(err.message).to.equal(`Specified address ${invoiceAccountAddress.addressId} is not in company ${companyId}`);
-        });
-
-        test('the invoice account address is not saved', async () => {
-          await expect(invoiceAccountsService.createInvoiceAccountAddress(invoiceAccountAddress))
-            .to.reject(errors.ConflictingDataError);
-          expect(invoiceAccountAddressesRepo.create.called).to.equal(false);
-        });
-      });
-
-      experiment('and the posted address ID belongs to the licence holder company', () => {
-        test('no error is thrown', async () => {
-          const func = () => invoiceAccountsService.createInvoiceAccountAddress(invoiceAccountAddress);
-          expect(func()).to.not.reject();
-        });
-      });
-
-      experiment('and the posted contact ID does not belong to the licence holder company', () => {
-        beforeEach(async () => {
-          invoiceAccountAddress.contactId = uuid();
-
-          contactsRepo.findOneWithCompanies.resolves({
-            companyContacts: [{
-              companyId: 'some-other-id'
-            }]
-          });
-        });
-
-        test('a ConflictingDataError is thrown', async () => {
-          const err = await expect(invoiceAccountsService.createInvoiceAccountAddress(invoiceAccountAddress))
-            .to.reject(errors.ConflictingDataError);
-          expect(err.message).to.equal(`Specified contact ${invoiceAccountAddress.contactId} is not in the company ${companyId}`);
-        });
-
-        test('the invoice account address is not saved', async () => {
-          await expect(invoiceAccountsService.createInvoiceAccountAddress(invoiceAccountAddress))
-            .to.reject(errors.ConflictingDataError);
-          expect(invoiceAccountAddressesRepo.create.called).to.equal(false);
-        });
-      });
-
-      experiment('and the posted contact ID belongs to the licence holder company', () => {
-        beforeEach(async () => {
-          invoiceAccountAddress.contactId = uuid();
-        });
-
-        test('no error is thrown', async () => {
-          const func = () => invoiceAccountsService.createInvoiceAccountAddress(invoiceAccountAddress);
-          expect(func()).to.not.reject();
-        });
-      });
-    });
-
-    experiment('when there is an agent company', () => {
-      let invoiceAccountAddress;
-
-      beforeEach(async () => {
-        invoiceAccountAddress = {
-          invoiceAccountId: uuid(),
-          addressId: uuid(),
-          startDate: '2020-04-01',
-          agentCompanyId: uuid(),
-          contactId: null
-        };
-      });
-
-      experiment('and the posted address ID does not belong to the agent company', () => {
-        beforeEach(async () => {
-          addressesRepo.findOneWithCompanies.resolves({
-            companyAddresses: [{
-              companyId: 'some-other-id'
-            }]
-          });
-        });
-
-        test('a ConflictingDataError is thrown', async () => {
-          const err = await expect(invoiceAccountsService.createInvoiceAccountAddress(invoiceAccountAddress))
-            .to.reject(errors.ConflictingDataError);
-          expect(err.message).to.equal(`Specified address ${invoiceAccountAddress.addressId} is not in company ${invoiceAccountAddress.agentCompanyId}`);
-        });
-
-        test('the invoice account address is not saved', async () => {
-          await expect(invoiceAccountsService.createInvoiceAccountAddress(invoiceAccountAddress))
-            .to.reject(errors.ConflictingDataError);
-          expect(invoiceAccountAddressesRepo.create.called).to.equal(false);
-        });
-      });
-
-      experiment('and the posted address ID belongs to the agent company', () => {
-        beforeEach(async () => {
-          addressesRepo.findOneWithCompanies.resolves({
-            companyAddresses: [{
-              companyId: invoiceAccountAddress.agentCompanyId
-            }]
-          });
-        });
-
-        test('no error is thrown', async () => {
-          await invoiceAccountsService.createInvoiceAccountAddress(invoiceAccountAddress);
-        });
-      });
-
-      experiment('and the posted contact ID does not belong to the agent company', () => {
-        beforeEach(async () => {
-          invoiceAccountAddress.contactId = uuid();
-
-          addressesRepo.findOneWithCompanies.resolves({
-            companyAddresses: [{
-              companyId: invoiceAccountAddress.agentCompanyId
-            }]
-          });
-
-          contactsRepo.findOneWithCompanies.resolves({
-            companyContacts: [{
-              companyId: 'some-other-id'
-            }]
-          });
-        });
-
-        test('a ConflictingDataError is thrown', async () => {
-          const err = await expect(invoiceAccountsService.createInvoiceAccountAddress(invoiceAccountAddress))
-            .to.reject(errors.ConflictingDataError);
-          expect(err.message).to.equal(`Specified contact ${invoiceAccountAddress.contactId} is not in the company ${invoiceAccountAddress.agentCompanyId}`);
-        });
-
-        test('the invoice account address is not saved', async () => {
-          await expect(invoiceAccountsService.createInvoiceAccountAddress(invoiceAccountAddress))
-            .to.reject(errors.ConflictingDataError);
-          expect(invoiceAccountAddressesRepo.create.called).to.equal(false);
-        });
-      });
-
-      experiment('and the posted contact ID belongs to the agent company', () => {
-        beforeEach(async () => {
-          invoiceAccountAddress.contactId = uuid();
-
-          addressesRepo.findOneWithCompanies.resolves({
-            companyAddresses: [{
-              companyId: invoiceAccountAddress.agentCompanyId
-            }]
-          });
-
-          contactsRepo.findOneWithCompanies.resolves({
-            companyContacts: [{
-              companyId: invoiceAccountAddress.agentCompanyId
-            }]
-          });
-        });
-
-        test('no error is thrown', async () => {
-          const func = () => invoiceAccountsService.createInvoiceAccountAddress(invoiceAccountAddress);
-          expect(func()).to.not.reject();
-        });
-      });
-    });
-
-    experiment('when the invoice account address data is valid', async () => {
-      let result, invoiceAccountId, invoiceAccountAddress;
-
-      beforeEach(async () => {
-        invoiceAccountId = uuid();
-        invoiceAccountAddress = {
-          invoiceAccountId,
-          addressId: uuid(),
-          startDate: '2020-04-01',
-          agentCompanyId: null,
-          contactId: null
-        };
-
-        invoiceAccountAddressesRepo.create.resolves({
-          invoiceAccountAddressId: 'test-id',
-          ...invoiceAccountAddress
-        });
-      });
-
-      test('invoice account addresses are found', async () => {
-        await invoiceAccountsService.createInvoiceAccountAddress(invoiceAccountAddress);
-        expect(
-          invoiceAccountAddressesRepo.findAll.calledWith(invoiceAccountId)
-        ).to.be.true();
-      });
-
-      experiment('when there are no date conflicts', async () => {
-        beforeEach(async () => {
-          result = await invoiceAccountsService.createInvoiceAccountAddress(invoiceAccountAddress);
-        });
-
-        test('the invoice account address is saved via the repository', async () => {
-          expect(invoiceAccountAddressesRepo.create.called).to.equal(true);
-        });
-
-        test('the saved invoice account address is returned', async () => {
-          expect(result.invoiceAccountAddressId).to.equal('test-id');
-        });
-      });
-
-      experiment('when there are date conflicts', async () => {
-        beforeEach(() => {
-          invoiceAccountAddressesRepo.findAll.resolves([
-            { startDate: '2016-01-01', endDate: '2018-05-02' },
-            { startDate: '2018-05-03', endDate: null }]);
-        });
-
-        test('a ConflictingDataError is thrown', async () => {
-          const err = await expect(invoiceAccountsService.createInvoiceAccountAddress(invoiceAccountAddress))
-            .to.reject();
-
-          expect(err.name).to.equal('ConflictingDataError');
-          expect(err.message).to.equal('Existing invoice account address exists for date range');
-        });
-
-        test('the invoice account address is not saved', async () => {
-          try {
-            await invoiceAccountsService.createInvoiceAccountAddress(invoiceAccountAddress);
-          } catch (err) {
-          }
-          expect(invoiceAccountAddressesRepo.create.called).to.equal(false);
-        });
-      });
-
-      experiment('when there are no address for the invoice account', async () => {
-        beforeEach(async () => {
-          invoiceAccountAddressesRepo.findAll.resolves([]);
-
-          result = await invoiceAccountsService.createInvoiceAccountAddress(invoiceAccountAddress);
-        });
-
-        test('the invoice account address is saved via the repository', async () => {
-          expect(invoiceAccountAddressesRepo.create.called).to.equal(true);
-        });
-
-        test('the saved invoice account address is returned', async () => {
-          expect(result.invoiceAccountAddressId).to.equal('test-id');
-        });
-      });
-    });
-  });
-
   experiment('.deleteInvoiceAccount', () => {
     test('calls the deleteOne repo method', async () => {
       await invoiceAccountsService.deleteInvoiceAccount('test-invoice-account-id');
       expect(invoiceAccountsRepo.deleteOne.calledWith('test-invoice-account-id')).to.be.true();
-    });
-  });
-
-  experiment('.deleteInvoiceAccountAddress', () => {
-    test('calls the deleteOne repo method', async () => {
-      await invoiceAccountsService.deleteInvoiceAccountAddress('test-invoice-account-address-id');
-      expect(invoiceAccountAddressesRepo.deleteOne.calledWith('test-invoice-account-address-id')).to.be.true();
     });
   });
 });
